Rock_vs_Mine_Prediction_Project
This project is a binary classification task to predict whether a given object is a rock or a mine using a dataset of sonar signal frequencies. It includes exploratory data analysis, feature selection, and a machine learning model (logistic regression) implemented in Python. The project was developed using Jupyter Notebook and utilizes libraries such as Pandas, Numpy, and Scikit-learn.